techdebt: [PL-0]: Separate Contracts of ResourceGroup Module
