+++ 820-platform-service/contracts/resource-group-service/BUILD.bazel
@@ -8,10 +8,13 @@ java_library(
     visibility = ["//visibility:public"],
     deps = [
         "//:lombok",
+        "//942-enforcement-sdk:module",
+        "//960-ng-core-beans:module",
         "//960-persistence:module",
         "//970-ng-commons:module",
         "//980-commons:module",
         "//999-annotations:module",
+        "//clients/access-control/sdk:module",
         "@maven//:com_fasterxml_jackson_core_jackson_annotations",
         "@maven//:com_google_code_findbugs_annotations",
         "@maven//:com_google_guava_guava",
@@ -33,10 +36,13 @@ java_library(
     visibility = ["//visibility:public"],
     deps = [
         ":module",
+        "//942-enforcement-sdk:module",
+        "//960-ng-core-beans:module",
         "//970-ng-commons:module",
         "//980-commons:module",
         "//990-commons-test:module",
         "//999-annotations:module",
+        "//clients/access-control/sdk:module",
         "@maven//:junit_junit",
         "@maven//:org_apache_commons_commons_lang3",
         "@maven//:org_assertj_assertj_core",
