+++ 820-platform-service/contracts/resource-group-service/src/main/java/io/harness/resourcegroup/v2/remote/dto/ResourceGroupResponse.java
@@ -5,12 +5,11 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.resourcegroupclient.remote.v2;
+package io.harness.resourcegroup.v2.remote.dto;
 
 import static io.harness.annotations.dev.HarnessTeam.PL;
 
 import io.harness.annotations.dev.OwnedBy;
-import io.harness.resourcegroup.v2.remote.dto.ResourceGroupDTO;
 
 import io.swagger.annotations.ApiModel;
 import io.swagger.v3.oas.annotations.media.Schema;
