--- 820-platform-service/contracts/resource-group-service/src/main/java/io/harness/serializer/morphia/v2/ResourceGroupBeansMorphiaRegistrar.java
@@ -1,26 +1,0 @@
-/*
- * Copyright 2021 Harness Inc. All rights reserved.
- * Use of this source code is governed by the PolyForm Shield 1.0.0 license
- * that can be found in the licenses directory at the root of this repository, also available at
- * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
- */
-
-package io.harness.serializer.morphia.v2;
-
-import io.harness.morphia.MorphiaRegistrar;
-import io.harness.morphia.MorphiaRegistrarHelperPut;
-import io.harness.resourcegroup.v2.model.ResourceGroup;
-
-import java.util.Set;
-
-public class ResourceGroupBeansMorphiaRegistrar implements MorphiaRegistrar {
-  @Override
-  public void registerClasses(Set<Class> set) {
-    set.add(ResourceGroup.class);
-  }
-
-  @Override
-  public void registerImplementationClasses(MorphiaRegistrarHelperPut h, MorphiaRegistrarHelperPut w) {
-    // Nothing to register
-  }
-}
