+++ 820-platform-service/libs/resource-group-service/src/main/java/io/harness/resourcegroup/framework/v2/remote/resource/HarnessResourceGroupResourceImpl.java
@@ -1,0 +1,124 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Shield 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt.
+ */
+
+package io.harness.resourcegroup.framework.v2.remote.resource;
+
+import static io.harness.data.structure.EmptyPredicate.isEmpty;
+import static io.harness.ng.core.utils.NGUtils.verifyValuesNotChanged;
+import static io.harness.resourcegroup.ResourceGroupPermissions.DELETE_RESOURCEGROUP_PERMISSION;
+import static io.harness.resourcegroup.ResourceGroupPermissions.EDIT_RESOURCEGROUP_PERMISSION;
+import static io.harness.resourcegroup.ResourceGroupPermissions.VIEW_RESOURCEGROUP_PERMISSION;
+import static io.harness.resourcegroup.ResourceGroupResourceTypes.RESOURCE_GROUP;
+import static io.harness.resourcegroup.v1.remote.dto.ManagedFilter.NO_FILTER;
+import static io.harness.utils.PageUtils.getNGPageResponse;
+
+import io.harness.accesscontrol.AccountIdentifier;
+import io.harness.accesscontrol.NGAccessControlCheck;
+import io.harness.annotations.dev.HarnessTeam;
+import io.harness.annotations.dev.OwnedBy;
+import io.harness.beans.Scope;
+import io.harness.beans.ScopeLevel;
+import io.harness.enforcement.client.annotation.FeatureRestrictionCheck;
+import io.harness.enforcement.constants.FeatureRestrictionName;
+import io.harness.ng.beans.PageRequest;
+import io.harness.ng.beans.PageResponse;
+import io.harness.ng.core.dto.ResponseDTO;
+import io.harness.resourcegroup.framework.v2.service.ResourceGroupService;
+import io.harness.resourcegroup.framework.v2.service.impl.ResourceGroupValidatorImpl;
+import io.harness.resourcegroup.v1.remote.dto.ManagedFilter;
+import io.harness.resourcegroup.v1.remote.dto.ResourceGroupFilterDTO;
+import io.harness.resourcegroup.v2.remote.dto.ResourceGroupDTO;
+import io.harness.resourcegroup.v2.remote.dto.ResourceGroupRequest;
+import io.harness.resourcegroup.v2.remote.dto.ResourceGroupResponse;
+import io.harness.security.annotations.InternalApi;
+import io.harness.security.annotations.NextGenManagerAuth;
+
+import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
+import com.google.inject.Inject;
+import java.util.Optional;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import org.apache.commons.lang3.tuple.Pair;
+
+@AllArgsConstructor(access = AccessLevel.PUBLIC, onConstructor = @__({ @Inject }))
+@NextGenManagerAuth
+@OwnedBy(HarnessTeam.PL)
+public class HarnessResourceGroupResourceImpl implements HarnessResourceGroupResource {
+  ResourceGroupService resourceGroupService;
+  ResourceGroupValidatorImpl resourceGroupValidator;
+
+  @NGAccessControlCheck(resourceType = RESOURCE_GROUP, permission = VIEW_RESOURCEGROUP_PERMISSION)
+  public ResponseDTO<ResourceGroupResponse> get(
+      String identifier, String accountIdentifier, String orgIdentifier, String projectIdentifier) {
+    Optional<ResourceGroupResponse> resourceGroupResponseOpt = Optional.ofNullable(
+        resourceGroupService.get(Scope.of(accountIdentifier, orgIdentifier, projectIdentifier), identifier, NO_FILTER)
+            .orElse(null));
+    return ResponseDTO.newResponse(resourceGroupResponseOpt.orElse(null));
+  }
+
+  @InternalApi
+  @NGAccessControlCheck(resourceType = RESOURCE_GROUP, permission = VIEW_RESOURCEGROUP_PERMISSION)
+  public ResponseDTO<ResourceGroupResponse> getInternal(
+      String identifier, String accountIdentifier, String orgIdentifier, String projectIdentifier) {
+    Optional<ResourceGroupResponse> resourceGroupResponseOpt =
+        Optional.ofNullable(resourceGroupService
+                                .get(Scope.of(accountIdentifier, orgIdentifier, projectIdentifier), identifier,
+                                    isEmpty(accountIdentifier) ? ManagedFilter.ONLY_MANAGED : ManagedFilter.NO_FILTER)
+                                .orElse(null));
+    return ResponseDTO.newResponse(resourceGroupResponseOpt.orElse(null));
+  }
+
+  public ResponseDTO<PageResponse<ResourceGroupResponse>> list(String accountIdentifier, String orgIdentifier,
+      String projectIdentifier, String searchTerm, PageRequest pageRequest) {
+    return ResponseDTO.newResponse(getNGPageResponse(resourceGroupService.list(
+        Scope.of(accountIdentifier, orgIdentifier, projectIdentifier), pageRequest, searchTerm)));
+  }
+
+  public ResponseDTO<PageResponse<ResourceGroupResponse>> list(
+      ResourceGroupFilterDTO resourceGroupFilterDTO, String accountIdentifier, PageRequest pageRequest) {
+    return ResponseDTO.newResponse(getNGPageResponse(resourceGroupService.list(resourceGroupFilterDTO, pageRequest)));
+  }
+
+  @NGAccessControlCheck(resourceType = RESOURCE_GROUP, permission = EDIT_RESOURCEGROUP_PERMISSION)
+  @FeatureRestrictionCheck(FeatureRestrictionName.CUSTOM_RESOURCE_GROUPS)
+  public ResponseDTO<ResourceGroupResponse> create(@AccountIdentifier String accountIdentifier, String orgIdentifier,
+      String projectIdentifier, ResourceGroupRequest resourceGroupRequest) {
+    resourceGroupRequest.getResourceGroup().setAllowedScopeLevels(
+        Sets.newHashSet(ScopeLevel.of(accountIdentifier, orgIdentifier, projectIdentifier).toString().toLowerCase()));
+    resourceGroupValidator.validateResourceGroup(resourceGroupRequest);
+    ResourceGroupResponse resourceGroupResponse =
+        resourceGroupService.create(resourceGroupRequest.getResourceGroup(), false);
+    return ResponseDTO.newResponse(resourceGroupResponse);
+  }
+
+  @NGAccessControlCheck(resourceType = RESOURCE_GROUP, permission = EDIT_RESOURCEGROUP_PERMISSION)
+  public ResponseDTO<ResourceGroupResponse> update(String identifier, String accountIdentifier, String orgIdentifier,
+      String projectIdentifier, ResourceGroupRequest resourceGroupRequest) {
+    resourceGroupRequest.getResourceGroup().setAllowedScopeLevels(
+        Sets.newHashSet(ScopeLevel.of(accountIdentifier, orgIdentifier, projectIdentifier).toString().toLowerCase()));
+    resourceGroupValidator.validateResourceGroup(resourceGroupRequest);
+    validateRequest(accountIdentifier, orgIdentifier, projectIdentifier, resourceGroupRequest.getResourceGroup());
+    return ResponseDTO.newResponse(
+        (resourceGroupService.update(resourceGroupRequest.getResourceGroup(), false)).orElse(null));
+  }
+
+  @NGAccessControlCheck(resourceType = RESOURCE_GROUP, permission = DELETE_RESOURCEGROUP_PERMISSION)
+  public ResponseDTO<Boolean> delete(
+      String identifier, String accountIdentifier, String orgIdentifier, String projectIdentifier) {
+    return ResponseDTO.newResponse(
+        resourceGroupService.delete(Scope.of(accountIdentifier, orgIdentifier, projectIdentifier), identifier));
+  }
+
+  private void validateRequest(
+      String accountIdentifier, String orgIdentifier, String identifier, ResourceGroupDTO resourceGroupDTO) {
+    verifyValuesNotChanged(
+        Lists.newArrayList(Pair.of(accountIdentifier, resourceGroupDTO.getAccountIdentifier())), false);
+    verifyValuesNotChanged(Lists.newArrayList(Pair.of(orgIdentifier, resourceGroupDTO.getOrgIdentifier())), false);
+    verifyValuesNotChanged(Lists.newArrayList(Pair.of(identifier, resourceGroupDTO.getIdentifier())), false);
+  }
+}
