+++ 820-platform-service/libs/resource-group-service/src/main/java/io/harness/resourcegroup/resourceclient/resourcegroup/ResourceGroupResourceImpl.java
@@ -27,8 +27,8 @@ import io.harness.resourcegroup.beans.ValidatorType;
 import io.harness.resourcegroup.framework.v1.service.Resource;
 import io.harness.resourcegroup.framework.v1.service.ResourceInfo;
 import io.harness.resourcegroup.v1.remote.dto.ResourceGroupFilterDTO;
+import io.harness.resourcegroup.v1.remote.dto.ResourceGroupResponse;
 import io.harness.resourcegroupclient.remote.ResourceGroupClient;
-import io.harness.resourcegroupclient.remote.v1.ResourceGroupResponse;
 
 import com.google.common.collect.ImmutableMap;
 import com.google.inject.Inject;
