+++ 820-platform-service/libs/resource-group-service/src/main/java/io/harness/resourcegroup/v1/model/ResourceGroup.java
