+++ 820-platform-service/libs/resource-group-service/src/main/java/io/harness/resourcegroup/v2/model/ResourceGroup.java
