+++ 930-ng-core-clients/src/main/java/io/harness/resourcegroupclient/remote/ResourceGroupClient.java
@@ -14,7 +14,7 @@ import io.harness.annotations.dev.OwnedBy;
 import io.harness.ng.beans.PageResponse;
 import io.harness.ng.core.dto.ResponseDTO;
 import io.harness.resourcegroup.v1.remote.dto.ResourceGroupFilterDTO;
-import io.harness.resourcegroupclient.remote.v1.ResourceGroupResponse;
+import io.harness.resourcegroup.v1.remote.dto.ResourceGroupResponse;
 
 import retrofit2.Call;
 import retrofit2.http.Body;
