+++ access-control/service/src/test/java/io/harness/accesscontrol/resources/resourcegroups/HarnessResourceGroupServiceImplTest.java
@@ -28,7 +28,7 @@ import io.harness.category.element.UnitTests;
 import io.harness.ng.core.dto.ResponseDTO;
 import io.harness.resourcegroup.v1.remote.dto.ResourceGroupDTO;
 import io.harness.resourcegroupclient.remote.ResourceGroupClient;
-import io.harness.resourcegroupclient.remote.v1.ResourceGroupResponse;
+import io.harness.resourcegroup.v1.remote.dto.ResourceGroupResponse;
 import io.harness.rule.Owner;
 
 import java.io.IOException;
